componentezando e separando a store em modulos
