+++ store/modules/pontos/index.js
@@ -1,0 +1,9 @@
+export default {
+  namespaced: true,
+  state: {
+
+  },
+  mutations: {},
+  actions: {},
+  getters: {}
+}
